+++ AutomatedLab/AutomatedLabAzure.psm1
@@ -1625,6 +1625,45 @@ function Get-LabAzureAvailableSku
     Group-Object -Property Skus, Offer |
     ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
 
+    # Linux
+    # Ubuntu - official
+    $publishers |
+    Where-Object PublisherName -eq 'Canonical' |
+    Get-AzVMImageOffer |
+    Where-Object Offer -match '0001-com-ubuntu-server-\w+$' |
+    Get-AzVMImageSku |
+    Where-Object Skus -notmatch 'arm64' |
+    Get-AzVMImage |
+    Group-Object -Property Skus, Offer |
+    ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
+    # RedHat - official
+    $publishers |
+    Where-Object PublisherName -eq 'RedHat' |
+    Get-AzVMImageOffer |
+    Where-Object Offer -eq 'RHEL' |
+    Get-AzVMImageSku |
+    Where-Object Skus -notmatch '(RAW|LVM|CI)' |
+    Get-AzVMImage |
+    Group-Object -Property Skus, Offer |
+    ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
+    # CentOS - Roguewave, sounds slightly suspicious
+    $publishers |
+    Where-Object PublisherName -eq 'OpenLogic' |
+    Get-AzVMImageOffer |
+    Where-Object Offer -eq CentOS |
+    Get-AzVMImageSku |
+    Get-AzVMImage |
+    Group-Object -Property Skus, Offer |
+    ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
+    # Kali
+    $publishers |
+    Where-Object PublisherName -eq 'Kali-Linux' |
+    Get-AzVMImageOffer |
+    Get-AzVMImageSku |
+    Get-AzVMImage |
+    Group-Object -Property Skus, Offer |
+    ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
+
     # Desktop
     $publishers |
     Where-Object PublisherName -eq 'MicrosoftWindowsDesktop' |