+++ AutomatedLab/AutomatedLabDiskImageWindows.psm1
@@ -104,7 +104,7 @@
 
     # RHEL, CentOS, Fedora, OpenSuse Tumbleweed et al
     $rhelPath = "$DriveLetter`:\.treeinfo" # TreeInfo Syntax https://release-engineering.github.io/productmd/treeinfo-1.0.html
-    $rhelPackageInfo = "$DriveLetter`:{0}\repodata"
+    $rhelPackageInfo = "$DriveLetter`:{0}\*\repodata"
     if (Test-Path -Path $rhelPath -PathType Leaf)
     {
         $contentMatch = (Get-Content -Path $rhelPath -Raw) -match '(?s)(?<=\[general\]).*?(?=\[)'
@@ -138,7 +138,7 @@
         if (-not $packageXml)
         {
             # CentOS ISO for some reason contained only GUIDs
-            $packageXml = Get-ChildItem -Path $rhelPackageInfo -PipelineVariable file -File |
+            $packageXml = Get-ChildItem -Path $rhelPackageInfo -ErrorAction SilentlyContinue -PipelineVariable file -File |
             Get-Content -TotalCount 10 |
             Where-Object { $_ -like "*<comps>*" } |
             ForEach-Object { $file.FullName } |
@@ -169,20 +169,30 @@
         $os
     }
 
-    # Ubuntu 2004+
+    # Ubuntu 2004+, Kali
     $ubuntuPath = "$DriveLetter`:\.disk\info"
     $ubuntuPackageInfo = "$DriveLetter`:\pool\main"
     if (Test-Path -Path $ubuntuPath -PathType Leaf)
     {
         $infoContent = Get-Content -Path $ubuntuPath -TotalCount 1
-        $null = $infoContent -match '(?:Ubuntu)(?:-Server)?\s+(?<Version>\d\d\.\d\d).*Release\s(?<Arch>i386|amd64)\s\((?<ReleaseDate>\d{8})'
-        $osversion = $Matches.Version
-        if (([version]$osversion) -lt '20.4')
+        if ($infoContent -like 'Kali*')
         {
-            Write-ScreenInfo -Type Error -Message "Skipping $IsoFile, AutomatedLab was only tested with 20.04 and newer."
+            $null = $infoContent -match '(?:Kali GNU\/Linux)?\s+(?<Version>\d\d\d\d\.\d).*\s+"(?<Name>[\w-]+)".*Official\s(?<Arch>i386|amd64).*(?<ReleaseDate>\d{8})'
+            $osversion = $Matches.Version
+            $name = 'Kali Linux {0}' -f $osversion
         }
+        else
+        {
+            $null = $infoContent -match '(?:Ubuntu)(?:-Server)?\s+(?<Version>\d\d\.\d\d).*Release\s(?<Arch>i386|amd64)\s\((?<ReleaseDate>\d{8})'
+            $osversion = $Matches.Version
+            $name = ($infoContent -split '\s-\s')[0]
+            if (([version]$osversion) -lt '20.4')
+            {
+                Write-ScreenInfo -Type Error -Message "Skipping $IsoFile, AutomatedLab was only tested with 20.04 and newer."
+            }
+        }
+
         $osDate = $Matches.ReleaseDate
-        $name = ($infoContent -split '\s-\s')[0]
 
         $os = New-Object -TypeName AutomatedLab.OperatingSystem($name, $IsoFile.FullName)
         if ($Matches.Arch -eq 'i386')